CRUD FINALIZADO / LISTAGEM DE PRODUTOS COM PROBLEMA NO CSS / CARRINHO INICIADO

+++ src/main/java/br/com/wda/OpenBeerProject/Repository/CervejaRepository.java
@@ -1,8 +1,3 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package br.com.wda.OpenBeerProject.Repository;
 
 import br.com.wda.OpenBeerProject.Entity.Cerveja;
@@ -18,7 +13,7 @@ import org.springframework.stereotype.Repository;
  *
  */
 @Repository
-public interface CervejaRepository extends JpaRepository<Cerveja, Long> {
+public interface CervejaRepository extends JpaRepository<Cerveja, Integer> {
     
      public Optional<Cerveja> findByNome(String nome);
